add the free_time_count function
